First pass at getting natural notes into the form
I'll keep working on better form patterns here. It is tricky - I feel
like by nature the form itself should have a mostly "flat" schema. Only
after the user completes the form and it is valid should we build it
into the final shape. This approach doesn't seem like it will scale well
though.

Another thought I have is to make a component for the union. In this
case, a component that produces a "note pool" via onChange. I can see
how this would work from a pure data stand point, but I am not sure how
the validation would plug into this. If we could figure this out though,
I think it would scale much nicer. The every union could be encapsulated
and the "main" schema would just receive an instance of the union.

Either way, this works for now