Got a pretty solid metronome going
I may want to merge this as is, but regardless the next step I think is
to extract the metronome work from the actual display. I want to come up
with a data structure that represents the current screen state. This is
basically the config with a few added props. Then we can have a
component just take in that structure and be display only.

So split what I have into 2 components:

- 1 with metronome, audio logic, and builds the data structure
- 2 that consumes the data structure and then renders